+++ apps/dashboard/src/lib/components/Pathways/components/PathwayLandingPage/components/Editor/HeaderForm.svelte
@@ -38,6 +38,7 @@
   type="text"
   bind:value={pathway.landingpage.videoUrl}
 />
+
 <div class="mt-7">
   <p class="font-bold mb-3">{$t('course.navItem.landing_page.editor.header_form.replace_cover')}</p>
   <PrimaryButton
