+++ apps/dashboard/src/lib/components/UploadWidget/index.svelte
@@ -95,6 +95,7 @@
   }
 
   onMount(handleSubmit);
+  $: console.log('sup', $handleOpenWidget.open);
 </script>
 
 <Modal
