+++ apps/dashboard/src/lib/utils/services/pathways/index.ts
@@ -60,7 +60,7 @@ const ID_QUERY = `
       banner_image,
       created_at,
       lesson (
-        is_complete
+        lesson_completion(id, profile_id, is_complete)
       ),
       group_id (
         groupmember (
@@ -100,6 +100,37 @@ export async function fetchPathway(pathwayId?: Pathway['id'], slug?: Pathway['sl
     data,
     error
   };
+}
+
+export async function updatePathway (
+  pathwayId: Pathway['id'],
+  avatar: string | undefined,
+  pathway: Partial<Pathway>
+) {
+  if (avatar && pathwayId) {
+    const filename = `pathway/${pathwayId + Date.now()}.webp`;
+
+    const { data } = await supabase.storage.from('avatars').upload(filename, avatar, {
+      cacheControl: '3600',
+      upsert: false
+    });
+
+    if (data) {
+      const { data: response } = supabase.storage.from('avatars').getPublicUrl(filename);
+
+      if (!response.publicUrl) return;
+
+      pathway.logo = response.publicUrl;
+    }
+  }
+
+  await supabase.from('pathway').update(pathway).match({ id: pathwayId });
+
+  return pathway.logo;
+}
+
+export async function deletePathway(pathwayId: Pathway['id']) {
+  return await supabase.from('pathway').update({ status: 'DELETED' }).match({ id: pathwayId });
 }
 
 export function addPathwayCourse(pathwayId: Pathway['id'], courseId: PathwayCourse['course_id']) {
