+++ apps/dashboard/src/routes/pathways/[id]/settings/+page.svelte
@@ -12,10 +12,15 @@
     Loading
   } from 'carbon-components-svelte';
   import TrashCan from 'carbon-icons-svelte/lib/TrashCan.svelte';
+
+  import type { Pathway } from '$lib/utils/types';
   import { t } from '$lib/utils/functions/translations';
-  import { pathway, pathwaySettings, RADIO_VALUE } from '$lib/components/Pathways/store';
-  import { deletePathway, updatePathways } from '$lib/utils/services/pathways';
+  import { snackbar } from '$lib/components/Snackbar/store';
+  import generateSlug from '$lib/utils/functions/generateSlug';
+  import { handleOpenWidget } from '$lib/components/CourseLandingPage/store';
+  import { deletePathway, updatePathway } from '$lib/utils/services/pathways';
   import { currentOrgDomain, currentOrg, currentOrgPath } from '$lib/utils/store/org';
+  import { pathway, pathwaySettings, RADIO_VALUE } from '$lib/components/Pathways/store';
 
   import PageNav from '$lib/components/PageNav/index.svelte';
   import PageBody from '$lib/components/PageBody/index.svelte';
@@ -25,12 +30,8 @@
   import { VARIANTS } from '$lib/components/PrimaryButton/constants';
   import UploadWidget from '$lib/components/UploadWidget/index.svelte';
   import PrimaryButton from '$lib/components/PrimaryButton/index.svelte';
-  import { handleOpenWidget } from '$lib/components/CourseLandingPage/store';
-  import PathwayContainer from '$lib/components/Pathways/components/PathwayContainer.svelte';
   import SectionTitle from '$lib/components/Org/SectionTitle.svelte';
-  import generateSlug from '$lib/utils/functions/generateSlug';
-  import type { Pathway } from '$lib/utils/types';
-  import { snackbar } from '$lib/components/Snackbar/store';
+  import PathwayContainer from '$lib/components/Pathways/components/PathwayContainer.svelte';
 
   let isSaving = false;
   let isDeleting = false;
@@ -74,7 +75,7 @@
         courses_certificate
       } = $pathwaySettings;
 
-      await updatePathways($pathway.id, avatar, {
+      await updatePathway($pathway.id, avatar, {
         title,
         logo,
         description,
